Petites modifs, cas extremes du path planning à gérer

+++ userFiles/ctrl/groups_ctrl/gr18/path/path_planning_gr18.h
@@ -21,7 +21,7 @@ typedef struct PositionXY
 /// path-planning main structure
 typedef struct PathPlanning
 {
-	float **map;
+	int **map;
 	int flag_trajectory;
 	int **list_goal;
 	int nb_goals;
